app/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
pika==1.2.0
psycopg2-binary==2.9.3
requests==2.27.1
six==1.16.0
spotipy==2.19.0
urllib3==1.26.8
